style: 패키지 재구성 및 reformat code
