+++ src/main/java/wooteco/subway/domain/line/LineSeries.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.line;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,7 +19,9 @@ public class LineSeries {
         lines.stream()
             .filter(line -> line.getId() == null)
             .findAny()
-            .ifPresent(line -> {throw new IdMissingException(line.getName() + " 노선에 ID가 없습니다.");});
+            .ifPresent(line -> {
+                throw new IdMissingException(line.getName() + " 노선에 ID가 없습니다.");
+            });
     }
 
     public void add(Line line) {
