+++ src/main/java/wooteco/subway/domain/property/Distance.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.property;
 
 import java.util.Objects;
