+++ src/main/java/wooteco/subway/dto/request/SectionRequest.java
@@ -1,18 +1,17 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.request;
 
-public class LineRequest {
-    private String name;
-    private String color;
+public class SectionRequest {
     private Long upStationId;
     private Long downStationId;
     private int distance;
 
-    public String getName() {
-        return name;
+    public SectionRequest() {
     }
 
-    public String getColor() {
-        return color;
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
     }
 
     public Long getUpStationId() {
