+++ src/main/java/wooteco/subway/entity/StationEntity.java
@@ -1,6 +1,6 @@
 package wooteco.subway.entity;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public class StationEntity {
