+++ src/main/java/wooteco/subway/ui/ExceptionAdviser.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import wooteco.subway.dto.ExceptionResponse;
+import wooteco.subway.dto.response.ExceptionResponse;
 import wooteco.subway.exception.CustomException;
 
 @RestControllerAdvice
