+++ src/main/java/wooteco/subway/ui/SectionController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import wooteco.subway.dto.SectionRequest;
+import wooteco.subway.dto.request.SectionRequest;
 import wooteco.subway.service.SectionService;
 
 @RestController
