+++ src/main/java/wooteco/subway/util/SimpleReflectionUtils.java
@@ -6,7 +6,6 @@ import java.util.Arrays;
 import org.springframework.util.ReflectionUtils;
 
 import wooteco.subway.exception.IdFieldNotFoundException;
-import wooteco.subway.exception.IdMissingException;
 
 public class SimpleReflectionUtils {
