+++ src/test/java/wooteco/subway/domain/ColorTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import wooteco.subway.domain.property.Color;
 import wooteco.subway.exception.InvalidRequestException;
 
 class ColorTest {
