+++ src/test/java/wooteco/subway/domain/DistanceTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import wooteco.subway.domain.property.Distance;
 import wooteco.subway.exception.InvalidRequestException;
 
 class DistanceTest {
