+++ src/test/java/wooteco/subway/domain/StationTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import wooteco.subway.domain.station.Station;
+
 class StationTest {
 
     @Test
