+++ src/test/java/wooteco/subway/domain/fixture/LineFixture.java
@@ -2,7 +2,7 @@ package wooteco.subway.domain.fixture;
 
 import java.util.List;
 
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 public class LineFixture {
