+++ src/test/java/wooteco/subway/util/SimpleReflectionUtilsTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 class SimpleReflectionUtilsTest {
